NlDBSearch/NlDBSearch/search_indexes.py
import datetime
from haystack import indexes
from NlDBSearch.models import Entry


class NoteIndex(indexes.SearchIndex, indexes.Indexable):
    text = indexes.CharField(document=True, use_template=True)

    def get_model(self):
        return Entry

    def index_queryset(self, using=None):
        """Used when the entire index for model is updated."""
        return self.get_model().objects.all()