requirements.txt
django-haystack
pyelasticsearch